+++ class15/PeonyF/AndroidArchitectureStudy/app/src/main/java/com/architecture/androidarchitecturestudy/ui/main/MainPresenter.kt
@@ -1,4 +1,4 @@
-package com.architecture.androidarchitecturestudy.ui
+package com.architecture.androidarchitecturestudy.ui.main
 
 import android.util.Log
 import com.architecture.androidarchitecturestudy.data.model.Movie
